backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# WebSocket client for OpenAI Realtime API
websockets==12.0

# Environment variables
python-dotenv==1.0.0

# CORS and HTTP
httpx==0.26.0
